ch(refactoring): - refactor stateless component
